Version 1.2.0

+++ admin/views/groups/view.html.php
@@ -27,12 +27,15 @@ class TZ_Portfolio_PlusViewGroups extends JViewLegacy
     protected $items        = null;
     protected $pagination   = null;
     protected $state        = null;
+    public $filterForm;
 
     function display($tpl = null){
 
-        $this -> items      = $this -> get('Items');
-        $this -> pagination = $this -> get('Pagination');
-        $this -> state      = $this -> get('State');
+        $this -> items          = $this -> get('Items');
+        $this -> state          = $this -> get('State');
+        $this -> pagination     = $this -> get('Pagination');
+        $this -> filterForm     = $this -> get('FilterForm');
+        $this -> activeFilters  = $this -> get('ActiveFilters');
 
         TZ_Portfolio_PlusHelper::addSubmenu('groups');
 
@@ -44,61 +47,44 @@ class TZ_Portfolio_PlusViewGroups extends JViewLegacy
     }
 
     protected function addToolbar(){
-        $doc    = JFactory::getDocument();
-        $bar    = JToolBar::getInstance();
+
+        $user   = TZ_Portfolio_PlusUser::getUser();
+
+        // Get the results for each action.
+        $canDo = TZ_Portfolio_PlusHelper::getActions('com_tz_portfolio_plus', 'group');
 
         JToolBarHelper::title(JText::_('COM_TZ_PORTFOLIO_PLUS_GROUP_FIELDS_MANAGER'), 'folder-3');
-        JToolBarHelper::addNew('group.add');
-        JToolBarHelper::editList('group.edit');
-        JToolBarHelper::publish('groups.publish', 'JTOOLBAR_PUBLISH', true);
-        JToolBarHelper::unpublish('groups.unpublish', 'JTOOLBAR_UNPUBLISH', true);
-        JToolBarHelper::deleteList(JText::_('COM_TZ_PORTFOLIO_PLUS_QUESTION_DELETE'),'groups.delete');
-        JToolBarHelper::divider();
-        JToolBarHelper::preferences('com_tz_portfolio_plus');
-        JToolBarHelper::divider();
 
-
-        // If the joomla is version 3.0
-        if(COM_TZ_PORTFOLIO_PLUS_JVERSION_COMPARE){
-            $doc -> addStyleSheet(JURI::base(true).'/components/com_tz_portfolio_plus/fonts/font-awesome-4.5.0/css/font-awesome.min.css');
+        if($canDo -> get('core.create') || (count($user->getAuthorisedFieldGroups('core.create'))) > 0 ) {
+            JToolBarHelper::addNew('group.add');
         }
 
-        $doc -> addStyleSheet(JURI::base(true).'/components/com_tz_portfolio_plus/css/style.min.css');
+        if ($canDo->get('core.edit' ) || $canDo -> get('core.edit.own')) {
+            JToolBarHelper::editList('group.edit');
+        }
 
-        JToolBarHelper::help('JHELP_CONTENT_ARTICLE_MANAGER',false,'http://wiki.templaza.com/TZ_Portfolio_Plus_v3:Administration#Group_Fields');
+        $canEditState   = $canDo->get('core.edit.state') || $canDo->get('core.edit.state.own');
+        $canDelete      = $canDo->get('core.delete') || $canDo->get('core.delete.own');
 
-        // Special HTML workaround to get send popup working
-        $docClass       = ' class="btn btn-small"';
-        $youtubeIcon    = '<i class="tz-icon-youtube tz-icon-14"></i>&nbsp;';
-        $wikiIcon       = '<i class="tz-icon-wikipedia tz-icon-14"></i>&nbsp;';
+        if ($canEditState) {
+            JToolBarHelper::publish('groups.publish', 'JTOOLBAR_PUBLISH', true);
+            JToolBarHelper::unpublish('groups.unpublish', 'JTOOLBAR_UNPUBLISH', true);
+            JToolbarHelper::checkin('groups.checkin');
+        }
 
-        $youtubeTitle   = JText::_('COM_TZ_PORTFOLIO_PLUS_VIDEO_TUTORIALS');
-        $wikiTitle      = JText::_('COM_TZ_PORTFOLIO_PLUS_WIKIPEDIA_TUTORIALS');
+        if ($canDelete) {
+            JToolBarHelper::deleteList(JText::_('COM_TZ_PORTFOLIO_PLUS_QUESTION_DELETE'), 'groups.delete');
+        }
 
-        $videoTutorial    ='<a'.$docClass.' onclick="Joomla.popupWindow(\'http://www.youtube.com/channel/UCykS6SX6L2GOI-n3IOPfTVQ/videos\', \''
-            .$youtubeTitle.'\', 800, 500, 1)"'.' href="#">'
-            .$youtubeIcon.$youtubeTitle.'</a>';
+        if ($user->authorise('core.admin', 'com_tz_portfolio_plus')
+            || $user->authorise('core.options', 'com_tz_portfolio_plus')) {
+            JToolBarHelper::preferences('com_tz_portfolio_plus');
+        }
 
-        $wikiTutorial    ='<a'.$docClass.' onclick="Joomla.popupWindow(\'http://wiki.templaza.com/Main_Page\', \''
-            .$wikiTitle.'\', 800, 500, 1)"'.' href="#">'
-            .$wikiIcon
-            .$wikiTitle.'</a>';
+        JToolBarHelper::help('JHELP_CONTENT_ARTICLE_MANAGER',false,
+            'https://www.tzportfolio.com/document/administration/29-how-to-use-group-fields-in-tz-portfolio-plus.html?tmpl=component');
 
-        $bar->appendButton('Custom',$videoTutorial,'youtube');
-        $bar->appendButton('Custom',$wikiTutorial,'wikipedia');
-
-
-        JHtmlSidebar::addFilter(
-            JText::_('JOPTION_SELECT_PUBLISHED'),
-            'filter_published',
-            JHtml::_('select.options',JHtml::_('jgrid.publishedOptions', array('archived' => false, 'trash' => false))
-                ,'value','text', $this->state->get('filter.published'), true)
-        );
-    }
-
-    protected function getSortFields()
-    {
-        return array('name' => JText::_('COM_TZ_PORTFOLIO_PLUS_NAME'),
-            'id' => JText::_('JGRID_HEADING_ID'));
+        TZ_Portfolio_PlusToolbarHelper::customHelp('https://www.youtube.com/channel/UCrLN8LMXTyTahwDKzQ-YOqg/videos'
+            ,'COM_TZ_PORTFOLIO_PLUS_VIDEO_TUTORIALS', 'youtube', 'youtube');
     }
 }
